Clarify event logging library boundaries
